Refactor getTaskHeading to utility
